db_api/requirements_mongo.txt
Flask
pandas
pymongo
pymongo[srv]
cassandra-driver